address gemini comments make common util for picker add debouncer to chat  attachment as well
